Parse correction (otto there is some info here for you)
This should fix all the issue I could think of for the parsing problems. Now will actually parse properly and return the parse to the queue. Otto this cause a small issue in the CentralServicesParser file. If you look in the ServiceNumber and NumberTickets files you can see how I changed the constructors which caused the error. You fix it really easily bu adding true or false as a second argument, I just don't know which one you will need for what you are doing. True usually means you are dealing with a default string for ServiceNumber and NumberTickets but just check what I wrote down.